practices/practice-04-langchain-agent/pyproject.toml
[project]
name = "practice-04-langchain-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.8",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
]
